refactor usage of addStandbyTransaction
