+++ src/jumpstart/saga.ts
@@ -121,20 +121,22 @@ export function* dispatchPendingERC20Transactions(
 
       yield* put(
         addStandbyTransaction({
-          __typename: 'TokenTransferV3',
-          type: TokenTransactionTypeV2.Received,
-          context: {
-            id: transactionHash,
+          transaction: {
+            __typename: 'TokenTransferV3',
+            type: TokenTransactionTypeV2.Received,
+            context: {
+              id: transactionHash,
+            },
+            transactionHash,
+            networkId,
+            amount: {
+              value,
+              tokenAddress,
+              tokenId,
+            },
+            address,
+            metadata: {},
           },
-          transactionHash,
-          networkId,
-          amount: {
-            value,
-            tokenAddress,
-            tokenId,
-          },
-          address,
-          metadata: {},
         })
       )
 
@@ -175,27 +177,29 @@ export function* dispatchPendingERC721Transactions(
 
         yield* put(
           addStandbyTransaction({
-            __typename: 'NftTransferV3',
-            type: TokenTransactionTypeV2.NftReceived,
-            context: {
-              id: transactionHash,
-            },
-            transactionHash,
-            networkId,
-            nfts: [
-              {
-                tokenId: tokenId.toString(),
-                contractAddress: contractAddress.toLowerCase(),
-                tokenUri,
-                metadata,
-                media: [
-                  {
-                    raw: metadata?.image,
-                    gateway: metadata?.image,
-                  },
-                ],
+            transaction: {
+              __typename: 'NftTransferV3',
+              type: TokenTransactionTypeV2.NftReceived,
+              context: {
+                id: transactionHash,
               },
-            ],
+              transactionHash,
+              networkId,
+              nfts: [
+                {
+                  tokenId: tokenId.toString(),
+                  contractAddress: contractAddress.toLowerCase(),
+                  tokenUri,
+                  metadata,
+                  media: [
+                    {
+                      raw: metadata?.image,
+                      gateway: metadata?.image,
+                    },
+                  ],
+                },
+              ],
+            },
           })
         )
