+++ src/viem/saga.test.ts
@@ -119,16 +119,20 @@ describe('sendPreparedTransactions', () => {
       .call(getViemWallet, networkConfig.viemChain.celo, false)
       .put(
         addStandbyTransaction({
-          ...mockStandbyTransactions[0],
-          feeCurrencyId: mockCeloTokenId,
-          transactionHash: '0xmockTxHash1',
+          transaction: {
+            ...mockStandbyTransactions[0],
+            feeCurrencyId: mockCeloTokenId,
+            transactionHash: '0xmockTxHash1',
+          },
         })
       )
       .put(
         addStandbyTransaction({
-          ...mockStandbyTransactions[1],
-          feeCurrencyId: mockCeloTokenId,
-          transactionHash: '0xmockTxHash2',
+          transaction: {
+            ...mockStandbyTransactions[1],
+            feeCurrencyId: mockCeloTokenId,
+            transactionHash: '0xmockTxHash2',
+          },
         })
       )
       .returns(['0xmockTxHash1', '0xmockTxHash2'])
